Rename submodule under `cupyx.scipy.fft`
