+++ cupyx/scipy/fft/__init__.py
@@ -1,6 +1,6 @@
 # flake8: NOQA
-from cupyx.scipy.fft.fft import *
-from cupyx.scipy.fft.fft import (
+from cupyx.scipy.fft._fft import *
+from cupyx.scipy.fft._fft import (
     __all__, __ua_domain__, __ua_convert__, __ua_function__)
-from cupyx.scipy.fft.fft import _scipy_150
+from cupyx.scipy.fft._fft import _scipy_150
 from cupyx.scipy.fft._helper import next_fast_len  # NOQA
